bwx-game.py:
Added import statments to work with cloud9, for some reason '*' doesnt work in the editor and this helps a little when writing code. this can be removed later

Renamed variables to be slightly more child/noob friendly

Tweaked the addition of Connections to match the workflow of Locations and other elements in the engine

In the process of commenting the code so new users can get a handle on some of the syntax

advent.py:

added World.add_connection() to match adding locations and other elements. This still needs some work as it required a change to Connection that is a bit messy and i rushed through it last night. the goal here is to try and make this feel like addin Locations and other elements and allow a developer to use a connection instance to do stuff to later (like some connection.hide(), etc)

moved some of the functions that were just floating to a Game static class. this isnt ideal, but i wanted to separate some of the engine-specifc stuff so it was clearer what was kid-created, and what was part of the enging. i think Game() shouldnot be static, and that creating a Game(), i.e. my_game = Game(), makes a lot of sense and may be needed later to control stuff outside the world, like inventory, players, etc